Move byte size conversion to separate file
